eembed_log to favor 64 bit
No real advantage in replacing "long" with "long long" in order to
get 64 bit on 32bit and smaller platforms.

This will mean that printing an int on a smaller platform may be more
expensive than the native "long" time, but not _much_ more expensive
for most platforms.

eembed-arduino.c: align types to log, but cast-narrow "int64_t" to
"long" for printing, with small potential for issues if anyone is
actually wanting to print large int64_t values from the arduino
firmware, which Serial.print does not support.