eembed_static_assert(expr, msg) c89 compat
Prior to C11, ISO C did not specify _Static_assert(expr, msg).

eembed_static_assert() provides a compatibility macro which will work
even if -std=c89 is declared, which will be needed until VisualStudio
versions which support C11 are in wide use.

Additionally, a macro eembed_static_assert_sizes_match(a,b) provides
a handy function for checking types.
Also, eembed_static_assert_equals(a,b) is provided.

Usage of
	eembed_static_assert_sizes_match(a,b);
which uses
	eembed_static_assert_equals(a,b)
which uses
	eembed_static_assert(expr, msg);
can be seen in
	tests/demo-eembed-static-assert.c
with
	gcc \
		-Isrc src/eembed.c tests/demo-eembed-static-assert.c
and
	gcc -DFAIL \
		-Isrc src/eembed.c tests/demo-eembed-static-assert.c
and
	gcc -std=c89 -DFAIL \
		-Isrc src/eembed.c tests/demo-eembed-static-assert.c